% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhattacharyya_coef.R
\name{bhattacharyyaC_D}
\alias{bhattacharyyaC_D}
\title{Bhattacharyya coefficient for two Dirichlet distributions}
\usage{
bhattacharyyaC_D(a1, a2)
}
\arguments{
\item{a1}{Vector of alpha parameters for Dirichlet distribution 1. All alpha must be > 0.}

\item{a2}{Vector of alpha parameters for Dirichlet distribution 2. All alpha must be > 0.}
}
\value{
The value of BC
}
\description{
Calculate the Bhattacharyya coefficient (BC) for two Dirichlet distributions
}
\examples{
high_agreement <- bhattacharyyaC_D(c(5, 4, 1), c(4, 4, 2));
low_agreement <- bhattacharyyaC_D(c(5, 4, 1), c(1, 5, 4));
}
