% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhattacharyya_coef.R
\name{bhattacharyyaC_MN}
\alias{bhattacharyyaC_MN}
\title{Bhattacharyya coefficient for two multinomial distributions}
\usage{
bhattacharyyaC_MN(p, q)
}
\arguments{
\item{p}{Vector of probabilities for multinomial distribution 1, should sum to 1.}

\item{q}{Vector of probabilities for multinomial distribution 2, should sum to 1.}
}
\value{
The value of BC
}
\description{
Calculate the Bhattacharyya coefficient (BC) for two multinomial distributions
}
\examples{
high_agreement <- bhattacharyyaC_MN(c(0.5, 0.4, 0.1), c(0.4, 0.4, 0.2));
low_agreement <- bhattacharyyaC_MN(c(0.5, 0.4, 0.1), c(0.1, 0.5, 0.4));
}
