% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bhattacharyya_coef.R
\name{bhattacharyyaC_D}
\alias{bhattacharyyaC_D}
\title{Bhattacharyya coefficient for two Dirichlet distributions}
\usage{
bhattacharyyaC_D(p, q)
}
\arguments{
\item{p}{Vector of mean probabilities for Dirichlet distribution 1, should sum to 1.}

\item{q}{Vector of mean probabilities for Dirichlet distribution 2, should sum to 1.}
}
\value{
The value of BC
}
\description{
Calculate the Bhattacharyya coefficient (BC) for two Dirichlet distributions
}
\examples{
high_agreement <- bhattacharyyaC_D(c(0.5, 0.4, 0.1), c(0.4, 0.4, 0.2));
low_agreement <- bhattacharyyaC_D(c(0.5, 0.4, 0.1), c(0.1, 0.5, 0.4));
}
